Fixed some minor problems

+++ ft_strncpy.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_strncpy.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: wmaguire <wmaguire@student.codam.nl>       +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/10/13 10:22:28 by keizerrijk        #+#    #+#             */
-/*   Updated: 2021/10/13 11:27:36 by wmaguire         ###   ########.fr       */
+/*                                                        ::::::::            */
+/*   ft_strncpy.c                                       :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: wmaguire <wmaguire@student.codam.nl>         +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2021/10/13 10:22:28 by keizerrijk    #+#    #+#                 */
+/*   Updated: 2021/10/20 16:18:32 by wmaguire      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strncpy(char *dest, char *src, unsigned int n)
+char	*ft_strncpy(char *dest, char *src, size_t n)
 {
-	unsigned int	iterator;
+	size_t	iterator;
 
 	iterator = 0;
 	while (iterator < n && src[iterator] != '\0')
